feat(events): move BaseEvent
